Split up components and adapt to backend changes
