+++ frontend/src/searchform/searchform.css
@@ -25,38 +25,6 @@
     font-weight: bold;
 }
 
-.ingredient-delete-button {
-    all: unset;
-    font-size: 10px;
-    font-weight: bold;
-    color: var(--primary);
-    border: 1px var(--warning) solid;
-    width: 10px;
-    height: 10px;
-    line-height: 5px;
-    padding: 2.5px;
-    text-align: center;
-    border-radius: 10px; 
-    cursor: pointer;
-    margin-top: 3px;
-    background-color: var(--warning);
-}
-
-.ingredient-container {
-    display: flex;
-    flex-direction: row;
-    max-width: 1024px;
-    border: 2px black ;
-    border-radius: 4px;;
-    margin: 1px 0px 1px 0px;
-    padding: 3px;
-    background-color: var(--secondary-highlight);
-}
-
-.ingredient-label {
-    flex-grow: 1;
-}
-
 .searchform-input-container {
     display: flex;
     flex-direction: row;
